+++ app_utils.py
@@ -2,7 +2,7 @@
 import sys
 import os
 
-APP_VERSION = "1.5.0"
+APP_VERSION = "1.6.0"
 
 
 def get_app_dir():